move BurnDetails to BurnAndRelease
